requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.1
pydantic==2.6.4
httpx==0.27.0
python-dotenv==1.0.1
openai==1.40.0
